Creating a CODEOWNERS FILE
Members who are responsible for specific areas of the codebase are known as code owners. CODEOWNERS is a specific file that defines and displays code owners. The CODEOWNERS file enables you find out who to contact if you have a query about specific paths or files in the repository, while the Responsibilities list offers you a broad picture of who's in charge of different areas of the project.